Refactor: move searchInputWithHeader to private scope
